requirements.txt
cryptography==1.8.1
cycler==0.10.0
decorator==4.0.11
distributed==1.16.3
dj-database-url==0.5.0
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.8.2
django-cors-headers==2.2.0
docutils==0.13.1
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
first==2.0.1
flake8==3.4.1
Flask==0.12.2
Flask-Cors==3.0.2
gast==0.2.0
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
lightgbm==2.0.12
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.10
MarkupSafe==0.23
mccabe==0.6.1
MechanicalSoup==0.9.0.post4
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.0
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pew==0.1.26
pickleshare==0.7.4
Pillow==5.1.0
pip-tools==1.9.0
pipenv==8.0.2
ply==3.10
progressbar2==3.34.3
prompt-toolkit==1.0.14
protobuf==3.5.1
psutil==5.2.2
psycopg2==2.7.4
py==1.4.33
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
PyMySQL==0.8.0
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
python-utils==2.2.0
pythonz-bd==1.11.4
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.18.4
resumable-urlretrieve==0.1.5
rope-py3k==0.9.4.post1
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.0
tblib==1.3.2
termcolor==1.1.0
testpath==0.3
toolz==0.8.2
tornado==4.5.1
tqdm==4.19.4
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
Werkzeug==0.13
whitenoise==3.3.1
widgetsnbextension==2.0.0
win-unicode-console==0.5
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
